requirements.txt
fastapi
uvicorn
langchain
langchain-openai
openai-whisper
pydantic
python-multipart